+++ src/api/index.ts
@@ -6,7 +6,12 @@ export {
   type UnauthorizedHandler,
 } from './http';
 export { syncApi } from './syncApi';
-export { remoteSyncApi } from './remoteSyncApi';
+export {
+  remoteSyncApi,
+  type RemoteSyncActionResponse,
+  type RemoteSyncRuntimeStatus,
+  type RemoteSyncSitePayload,
+} from './remoteSyncApi';
 export { mqttApi } from './mqttApi';
 export { siteConfigApi, type SiteConfig } from './siteConfigApi';
 export {
@@ -17,10 +22,6 @@ export {
 export {
   deploymentSitesApi,
   type DeploymentSiteSummary,
-  type DeploymentSiteCreatePayload,
-  type DeploymentSiteUpdatePayload,
-  type DeploymentSiteTask,
-  type DeploymentSiteHealthcheckResult,
 } from './deploymentSitesApi';
 export {
   adminAuthApi,